+++ Views/CharacterWorkshop/BossesSectionView.axaml.cs
@@ -1,0 +1,14 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace FujinTerm.Views.CharacterWorkshop;
+
+public partial class BossesSectionView : UserControl
+{
+    public BossesSectionView()
+    {
+        InitializeComponent();
+    }
+
+    private void InitializeComponent() => AvaloniaXamlLoader.Load(this);
+}